requirements.txt
torch>=1.13.0
torchvision
tensorboard==2.13.0
einops==0.6.1
gradio==3.35.2
matplotlib==3.7.1
numpy==1.24.1
opencv_python==4.8.0.74
pandas==2.0.2
Pillow==9.5.0
scipy==1.11.1
seaborn==0.12.2
tqdm==4.65.0
scikit_image==0.21.0
protobuf==3.20
fastapi==0.99.0
